Test compilation des fichiers .sh Changement du type de données Y Cb Cr en uint8_t

+++ include/conversion.h
@@ -1,11 +1,14 @@
 #ifndef _CONV_H
 #define _CONV_H
 
+#include <stdint.h>
+
 #include "../include/ppm.h"
 
 /* Définition et déclaration d'un triplet de couleurs YCbCr */
+/* Y is defined to have a range of 16 to 235; Cb and Cr are defined to have a range of 16 to 240*/
 typedef struct triplet_ycbcr {
-    float triplet[3];
+    uint8_t triplet[3];
 } triplet_ycbcr;
 
 /*
